Began full implementation of Worldspace
During this, we ran into issues involving memory which has yet to be fixed, being the reason for the creation of a new branch. One issue involves the access of Objects within worldspace. Specifically, many null Object pointers which do not appear null when checked appear within the Worldspace, and cause problems when updating the space. The other issue involves the program calling SIGABRT when a RenderDrawRectEvent is deleted.

Additionally, this commit also contains several changes to the game's framework which has yet to be pushed onto the framework's git